feat: :recycle: refactor global styles and theme
