+++ src/styles/themes/default.ts
@@ -1,6 +1,8 @@
 export const theme = {
   bg: "#59358c",
+  bgSec: "#e94d5f",
   yellow: "#f7a706",
+  blue: "#1c6ea4",
   white: "#fff",
-  bgSec: "#e94d5f"
+  black: "#000",
 };
